create ascii art class, separate word class, test classes in main
