+++ explosive.rb
@@ -1,0 +1,122 @@
+class Image
+
+def initialize
+  @bombs = [
+
+  def guess_eight
+    return "
+
+                        . . .
+                         ||/
+                       `--+--'
+                         /|
+                        ' | '
+                          |
+                          |
+                      ,--'#`--.
+                      |#######|
+                   _.-'#######`-._
+                ,-'###############`-.
+              ,'#####################`,
+             /#########################|
+            |###########################|
+           |#############################|
+           |#############################|
+           |#############################|
+           |#############################|
+            |###########################|
+             |#########################/
+              `.#####################,'
+                `._###############_,'
+                   `--..#####..--'
+"
+  end
+
+  def guess_seven
+    return "
+
+
+
+
+
+
+
+
+                      ,--'#`--.
+                      |#######|
+                   _.-'#######`-._
+                ,-'###############`-.
+              ,'#####################`,
+             /#########################|
+            |###########################|
+           |#############################|
+           |#############################|
+           |#############################|
+           |#############################|
+            |###########################|
+             |#########################/
+              `.#####################,'
+                `._###############_,'
+                   `--..#####..--'
+"
+  end
+
+  def guess_six
+    return "
+
+
+
+
+
+
+
+
+
+
+
+
+              ,'#####################`,
+             /#########################|
+            |###########################|
+           |#############################|
+           |#############################|
+           |#############################|
+           |#############################|
+            |###########################|
+             |#########################/
+              `.#####################,'
+                `._###############_,'
+                   `--..#####..--'
+"
+  end
+
+
+  def guess_six
+    return "
+
+
+
+
+
+
+
+
+
+
+
+
+              ,'#####################`,
+             /#########################|
+            |###########################|
+           |#############################|
+           |#############################|
+           |#############################|
+           |#############################|
+            |###########################|
+             |#########################/
+              `.#####################,'
+                `._###############_,'
+                   `--..#####..--'
+"
+  end
+end
